# api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/lines": {
      "post": {
        "operationId": "LinesController_create",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLineDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure or route integrity violation."
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a line in the current tenant.",
        "tags": [
          "Lines"
        ]
      },
      "get": {
        "operationId": "LinesController_list",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 25,
              "example": 25,
              "type": "number"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "directionMode",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "SINGLE",
                "BOTH"
              ]
            }
          },
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "OUTBOUND",
                "RETURN"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "central",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedLinesResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure."
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List lines in the current tenant.",
        "tags": [
          "Lines"
        ]
      }
    },
    "/lines/{id}": {
      "get": {
        "operationId": "LinesController_getById",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          },
          "404": {
            "description": "Line not found in current tenant."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a line detail by id in the current tenant.",
        "tags": [
          "Lines"
        ]
      },
      "patch": {
        "operationId": "LinesController_update",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLineDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure or route integrity violation."
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          },
          "404": {
            "description": "Line not found in current tenant."
          },
          "409": {
            "description": "The proposed line would break existing reservations.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WouldBreakReservationsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update line fields in the current tenant.",
        "tags": [
          "Lines"
        ]
      },
      "put": {
        "operationId": "LinesController_replace",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLineDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure or route integrity violation."
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          },
          "404": {
            "description": "Line not found in current tenant."
          },
          "409": {
            "description": "The proposed line would break existing reservations.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WouldBreakReservationsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Replace line fields in the current tenant.",
        "tags": [
          "Lines"
        ]
      },
      "delete": {
        "operationId": "LinesController_remove",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cascade",
            "required": false,
            "in": "query",
            "description": "When true, also deactivates related rides and cancels active reservations.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          },
          "404": {
            "description": "Line not found in current tenant."
          },
          "409": {
            "description": "Line has active rides and cascade override is not enabled."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a line in the current tenant.",
        "tags": [
          "Lines"
        ]
      }
    },
    "/lines/{id}/stops": {
      "put": {
        "operationId": "LinesController_replaceStops",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceLineStopsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure or route integrity violation."
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          },
          "404": {
            "description": "Line not found in current tenant."
          },
          "409": {
            "description": "The proposed stop sequence would break existing reservations.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WouldBreakReservationsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Replace ordered intermediate stops for a line in the current tenant.",
        "tags": [
          "Lines"
        ]
      }
    },
    "/lines/{id}/reverse": {
      "post": {
        "operationId": "LinesController_createReverse",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure or route integrity violation."
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          },
          "404": {
            "description": "Line not found in current tenant."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a reverse line from an existing line route in the current tenant.",
        "tags": [
          "Lines"
        ]
      }
    },
...
  },
  "info": {
    "title": "SURP API",
    "description": "Backend API for SURP, providing tenant-scoped authentication, operational health endpoints, and secure session management for transportation workflows. Protected endpoints require BOTH Authorization: Bearer <accessToken> and X-Tenant-Slug headers (except /platform/* routes), and the tenant must match the token claim.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token issued by /auth/login. Use together with X-Tenant-Slug on protected endpoints."
      }
    },
    "schemas": {
...
      "WouldBreakReservationsDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "WOULD_BREAK_RESERVATIONS"
          },
          "invariant": {
            "type": "string",
            "example": "reservation.seatWithinCapacity",
            "description": "Key of the invariant the change would violate."
          },
          "affectedCount": {
            "type": "number",
            "example": 8,
            "description": "New invariant violations the proposed change would introduce."
          },
          "message": {
            "type": "string",
            "example": "Ova izmena cini 8 rezervacija nevidljivim."
          }
        },
        "required": [
          "code",
          "invariant",
          "affectedCount",
          "message"
        ]
      },
      "LineStopInputDto": {
        "type": "object",
        "properties": {
          "stationId": {
            "type": "string",
            "example": "station-intermediate-id"
          },
          "orderIndex": {
            "type": "number",
            "example": 1,
            "minimum": 1
          },
          "isBoarding": {
            "type": "boolean",
            "example": true,
            "default": true,
            "description": "Whether passengers may board at this stop."
          },
          "isDropoff": {
            "type": "boolean",
            "example": true,
            "default": true,
            "description": "Whether passengers may get off at this stop."
          }
        },
        "required": [
          "stationId",
          "orderIndex"
        ]
      },
      "CreateLineDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Central - North"
          },
          "departureStationId": {
            "type": "string",
            "example": "station-departure-id"
          },
          "arrivalStationId": {
            "type": "string",
            "example": "station-arrival-id"
          },
          "directionMode": {
            "type": "string",
            "enum": [
              "SINGLE",
              "BOTH"
            ],
            "default": "BOTH"
          },
          "direction": {
            "type": "string",
            "enum": [
              "OUTBOUND",
              "RETURN"
            ],
            "default": "OUTBOUND"
          },
          "pairKey": {
            "type": "string",
            "example": "seed-station-main-seed-station-referenced-1"
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "default": true
          },
          "intermediateStops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LineStopInputDto"
            }
          }
        },
        "required": [
          "departureStationId",
          "arrivalStationId"
        ]
      },
      "LineStationSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "BUS_STATION",
              "BUS_STOP"
            ],
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "address",
          "category",
          "isActive"
        ]
      },
      "LineIntermediateStopDto": {
        "type": "object",
        "properties": {
          "stationId": {
            "type": "string"
          },
          "stationName": {
            "type": "string"
          },
          "orderIndex": {
            "type": "number",
            "example": 1
          },
          "isBoarding": {
            "type": "boolean",
            "example": true
          },
          "isDropoff": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "stationId",
          "stationName",
          "orderIndex",
          "isBoarding",
          "isDropoff"
        ]
      },
      "LineResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "createdById": {
            "type": "object",
            "nullable": true
          },
          "updatedById": {
            "type": "object",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "departureStationId": {
            "type": "string"
          },
          "arrivalStationId": {
            "type": "string"
          },
          "directionMode": {
            "type": "string",
            "enum": [
              "SINGLE",
              "BOTH"
            ]
          },
          "direction": {
            "type": "string",
            "enum": [
              "OUTBOUND",
              "RETURN"
            ]
          },
          "pairKey": {
            "type": "object",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "departureStation": {
            "$ref": "#/components/schemas/LineStationSummaryDto"
          },
          "arrivalStation": {
            "$ref": "#/components/schemas/LineStationSummaryDto"
          },
          "intermediateStops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LineIntermediateStopDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "createdById",
          "updatedById",
          "name",
          "departureStationId",
          "arrivalStationId",
          "directionMode",
          "direction",
          "pairKey",
          "isActive",
          "departureStation",
          "arrivalStation",
          "intermediateStops",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedLinesResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LineResponseDto"
            }
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 25
          },
          "total": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ]
      },
      "UpdateLineDto": {
        "type": "object",
        "properties": {
          "confirmBreakingChange": {
            "type": "boolean",
            "example": false,
            "description": "Confirms a change previously refused with WOULD_BREAK_RESERVATIONS."
          },
          "name": {
            "type": "string",
            "example": "Central - North Updated"
          },
          "departureStationId": {
            "type": "string",
            "example": "station-departure-id"
          },
          "arrivalStationId": {
            "type": "string",
            "example": "station-arrival-id"
          },
          "directionMode": {
            "type": "string",
            "enum": [
              "SINGLE",
              "BOTH"
            ]
          },
          "direction": {
            "type": "string",
            "enum": [
              "OUTBOUND",
              "RETURN"
            ]
          },
          "pairKey": {
            "type": "string",
            "example": "seed-station-main-seed-station-referenced-1"
          },
          "isActive": {
            "type": "boolean",
            "example": false
          },
          "intermediateStops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LineStopInputDto"
            }
          }
        }
      },
      "ReplaceLineStopsDto": {
        "type": "object",
        "properties": {
          "confirmBreakingChange": {
            "type": "boolean",
            "example": false,
            "description": "Confirms a change previously refused with WOULD_BREAK_RESERVATIONS."
          },
          "intermediateStops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LineStopInputDto"
            }
          }
        },
        "required": [
          "intermediateStops"
        ]
      },
...
    }
  }
}